refactor(cron): remove the core execution test proxy
Import executeJobCore from its existing production owner in three test
suites. Remove the redundant global member, forwarder and copied result
type while retaining the onTimer boundary and every test body.
