+++ src/cron/service.skips-main-jobs-empty-systemevent-text.test.ts
@@ -7,7 +7,7 @@ import {
   withCronServiceForTest,
 } from "./service.test-harness.js";
 import { createCronServiceState } from "./service/state.js";
-import { executeJobCore } from "./service/timer.test-support.js";
+import { executeJobCore } from "./service/timer-execution.js";
 import type { CronJob } from "./types.js";
 
 const noopLogger = createNoopLogger();
